fix: move health check to non abci skill
