+++ packages/valory/services/optimus/service.yaml
@@ -6,7 +6,7 @@ aea_version: '>=1.0.0, <2.0.0'
 license: Apache-2.0
 fingerprint: {}
 fingerprint_ignore_patterns: []
-agent: valory/optimus:0.1.0:bafybeieexv73fdbkae3qzi5ozunexl7r456v2go7u4nwj6dg4q4nzlf4he
+agent: valory/optimus:0.1.0:bafybeicoabd2bocbozel3q7cjhy5dkmqhrwjfgaqrvxlylspghiehfi5ru
 number_of_agents: 1
 deployment:
   agent:
