+++ packages/valory/skills/llm_interaction/behaviours.py
@@ -1,0 +1,26 @@
+# -*- coding: utf-8 -*-
+# ------------------------------------------------------------------------------
+#
+#   Copyright 2023-2024 Valory AG
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+#
+# ------------------------------------------------------------------------------
+
+"""This package contains the implementation of ."""
+from aea.skills.behaviours import SimpleBehaviour
+
+class LlmInteractionBehaviour(SimpleBehaviour):
+    def act(self) -> None:
+        """Implement the act."""
+        self.context.logger.info("INSIDE NON ABCI SKILL LLM INTERACTION")
