+++ packages/valory/skills/optimus_abci/models.py
@@ -74,7 +74,6 @@ class Params(  # pylint: disable=too-many-ancestors
 
     def __init__(self, *args: Any, **kwargs: Any) -> None:
         """Init"""
-        self.service_endpoint_base = self._ensure("service_endpoint_base", kwargs, str)
         super().__init__(*args, **kwargs)
