+++ pyproject.toml
@@ -40,3 +40,5 @@ pycoingecko = "==3.2.0"
 numpy = "==2.2.0"
 pandas = "^2.2.3"
 pyfolio = "^0.9.2"
+peewee = "==3.17.5"
+openai = "==1.55.3"
